auth-gateway/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
httpx>=0.25.0
